fix: long -> int 타입으로 변경
- update 메소드는 int 형을 반환한다

+++ src/main/java/com/prgrms/vouchermanagement/customer/repository/CustomerJdbcRepository.java
@@ -105,12 +105,12 @@ public class CustomerJdbcRepository implements CustomerRepository {
 	}
 
 	@Override
-	public long deleteAll() {
+	public int deleteAll() {
 		return jdbcTemplate.update("DELETE FROM customers");
 	}
 
 	@Override
-	public long deleteById(UUID customerId) {
+	public int deleteById(UUID customerId) {
 		return jdbcTemplate.update("DELETE FROM customers WHERE customer_id = UUID_TO_BIN(?)",
 			customerId.toString().getBytes());
 	}
